+++ NEWS.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+- Add {crosstalk} support for linked brushing and filtering with other crosstalk-enabled widgets (#95)
+- If `id` column is not provided in the data for a timevis, the row names will be used as IDs
+
 # timevis 2.1.0 (2022-10-22)
 
 - Add functions `zoomIn()` and `zoomOut()`